% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{corr}
\alias{corr}
\title{corr}
\usage{
corr(n, corr, mu = c(0, 0), sd = 1, m = 0, mode = "matrix")
}
\arguments{
\item{n}{data length}

\item{corr}{target correlation}

\item{mu}{data mean vector (for matrix mode)}

\item{sd}{standard deviation (for vector mode)}

\item{m}{mean (for vector mode)}

\item{mode}{either "matrix" or "vector"}
}
\description{
Generate correlated data with given parameters, either as a matrix or as a single vector.
}
