% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MC.R
\name{mc.corr.norm}
\alias{mc.corr.norm}
\title{mc.corr.norm}
\usage{
mc.corr.norm(r, n, m = 0, sd = 1)
}
\arguments{
\item{r}{target correlation}

\item{n}{length, default = 100}

\item{m}{mean, default = 0}

\item{sd}{standard deviation, default = 1}
}
\description{
Generated Monte Carlo simulated data with given normal distribution information and target correlation
}
