% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MC.R
\name{mc.corr}
\alias{mc.corr}
\title{mc.corr}
\usage{
mc.corr(n, corr, mu = c(0, 0))
}
\arguments{
\item{n}{data length}

\item{corr}{target correlation}

\item{mu}{data mean vector}
}
\description{
Generate 2 rows matrix that correlation is almost fit to target value
}
