% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/math.R
\name{mcnorm}
\alias{mcnorm}
\title{mcnorm}
\usage{
mcnorm(n = 100, s = 1, m = 0, z = 0)
}
\arguments{
\item{n}{length, default = 100}

\item{s}{standard deviation, default = 1}

\item{m}{mean, default = 0}

\item{z}{offset, default = 0}
}
\description{
Generated Monte Carlo simulated data that follow normal distribution
}
