% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.cat}
\alias{print.cat}
\title{print.cat}
\usage{
\method{print}{cat}(title, value)
}
\arguments{
\item{title}{String}

\item{value}{Any Type}
}
\description{
Print title and value using cat and print function
}
