% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.stat2}
\alias{print.stat2}
\title{print.stat2}
\usage{
\method{print}{stat2}(a, b)
}
\arguments{
\item{a}{Any Type}

\item{b}{Any Type}
}
\description{
Print statistic information and relationship of two data
}
