% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.summary}
\alias{print.summary}
\title{print.summary}
\usage{
\method{print}{summary}(x, title = "")
}
\arguments{
\item{x}{Value}
}
\description{
Print summary and data length, mean, variance, standard deviation, and CV
}
