% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.vector}
\alias{print.vector}
\title{print.vector}
\usage{
\method{print}{vector}(names_vec, vec)
}
\arguments{
\item{names_vec}{String Vector}

\item{vec}{Vector}
}
\description{
Print vector with given names string vector
}
