% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.vector.stat}
\alias{print.vector.stat}
\alias{print.data.frame.stat}
\title{print.vector.stat}
\usage{
\method{print}{data.frame.stat}(data)
}
\arguments{
\item{x}{data.frame}
}
\description{
Print data frame statistic information
}
